+++ pilot/model/BUILD
@@ -49,6 +49,7 @@ go_test(
     srcs = [
         "config_test.go",
         "conversion_test.go",
+        "egress_rules_test.go",
     ],
     deps = [
         ":go_default_library",
